backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.9.2
alembic==1.13.2
pytest==8.3.2
httpx==0.27.2
pytest-asyncio==0.23.8
